+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## 0.3.2 (unreleased)
 
+FEATURES:
+
+* New command: `packer inspect`. This command tells you the components of
+  a template. It respects the `-machine-readable` flag as well so you can
+  parse out components of a template.
+
 IMPROVEMENTS:
 
 * builder/virtualbox: Do not check for VirtualBox as part of template